job-scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
pandas==2.1.3
tqdm==4.66.1
playwright==1.40.0 